Introduce MarkdownStyleUtil
